# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlightGo Backend API",
    "version": "1.0.0",
    "description": "FlightGo Backend API"
  },
  "servers": [
    {
      "url": "undefined",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Airport": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "The airport code (3 uppercase letters)"
          },
          "city": {
            "type": "string",
            "description": "The city where the airport is located"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/airport": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new airport (Admin only)",
        "tags": [
          "Airport"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Airport created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get all airports (Admin only)",
        "tags": [
          "Airport"
        ],
        "responses": {
          "200": {
            "description": "List of airports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Airport"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          }
        }
      }
    },
    "/airport/{code}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get an existing airport (Admin only)",
        "tags": [
          "Airport"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "description": "The airport code (3 uppercase letters)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Airport found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "404": {
            "description": "Airport not found"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an existing airport (Admin only)",
        "tags": [
          "Airport"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "description": "The airport code (3 uppercase letters)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "description": "The new city of the airport"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Airport updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Airport not found"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete an airport (Admin only)",
        "tags": [
          "Airport"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "description": "The airport code (3 uppercase letters)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Airport deleted successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized, valid JWT token required"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Airport not found"
          }
        }
      }
    },
...
  },
...
}